xdr/requirements.txt
# XDR Production Dependencies
flask>=3.0
PyJWT>=2.8
PyYAML>=6.0
